Search books corrigido

+++ app/src/main/java/com/example/desafioaula2ioasys/data_remote/model/BooksListResponse.kt
@@ -11,6 +11,6 @@ data class BooksListResponse(
     @SerializedName("totalItems")
     val totalItems: Int,
     @SerializedName("totalPages")
-    val totalPages: Int
+    val totalPages: Double
 )
