+++ app/src/main/java/com/example/desafioaula2ioasys/domain/model/ListBooks.kt
@@ -1,10 +1,8 @@
 package com.example.desafioaula2ioasys.domain.model
 
-import com.example.desafioaula2ioasys.data_remote.model.BookResponse
-
 data class ListBooks(
     val data: MutableList<Book>,
     val page: Int,
     val totalItems: Int,
-    val totalPages: Int
+    val totalPages: Double
 )
